+++ packages/cli/src/main_parser_factory.ts
@@ -6,6 +6,7 @@ import { createPipelineDeployCommand } from './commands/pipeline-deploy/pipeline
 import { createConfigureCommand } from './commands/configure/configure_command_factory.js';
 import { generateCommandFailureHandler } from './error_handler.js';
 import { createInfoCommand } from './commands/info/info_command_factory.js';
+import * as path from 'path';
 
 /**
  * Creates main parser.
@@ -24,6 +25,9 @@ export const createMainParser = (
       description: 'Print debug logs to the console',
     })
     .strict()
+    // pnpm places the command bin file in `ampx.js` instead of `ampx` which causes yargs to think the command name is `ampx.js`.
+    // This tells yargs that the command name is `ampx`.
+    .scriptName(path.parse(process.argv[1]).name)
     .command(createGenerateCommand())
     .command(createSandboxCommand())
     .command(createPipelineDeployCommand())